Several skirmish and trait fixes. New traits. Release 0.95.093

+++ trunk/open-ig/src/hu/openig/model/FleetStatistics.java
@@ -26,9 +26,9 @@ public class FleetStatistics {
 	/** The maximum speed. */
 	public int speed;
 	/** The firepower. */
-	public int firepower;
+	public double firepower;
 	/** The ground firepower. */
-	public int groundFirepower;
+	public double groundFirepower;
 	/** The nearby planet, where distance < 45 units. */
 	public Planet planet;
 }
